LayoutTests/media/modern-media-controls/start-support/start-support-manual-play-expected.txt

Testing the StartSupport behavior with manual play.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Media is playing
PASS mediaController.controls.showsStartButton is false

Media is paused
PASS mediaController.controls.showsStartButton is false

PASS successfullyParsed is true

TEST COMPLETE
